Use preset.ts instead of index.ts
